update syscalls to use long type, update tests

+++ sched.h
@@ -1130,7 +1130,7 @@ struct task_struct {
 	/* Used by LSM modules for access restriction: */
 	void				*security;
 #endif
-	int weight;
+	long weight;
 	/*
 	 * New fields for task_struct should be added above here, so that
 	 * they are included in the randomized portion of task_struct.
