+++ syscalls.h
@@ -941,8 +941,8 @@ asmlinkage long sys_pkey_free(int pkey);
 asmlinkage long sys_statx(int dfd, const char __user *path, unsigned flags,
 			  unsigned mask, struct statx __user *buffer);
 asmlinkage long sys_hello(void);
-asmlinkage int sys_set_weight(int weight);
-asmlinkage int sys_get_total_weight(void);
+asmlinkage long sys_set_weight(int weight);
+asmlinkage long sys_get_total_weight(void);
 asmlinkage long sys_get_heaviest_child(void);
 
 #endif
